requirements.txt
PyQt5==5.15.7
mysql-connector-python==8.0.33